requirements.txt
kafka-python
pymongo
psycopg2-binary
streamlit
requests
pandas
sqlalchemy